app/requirements.txt
aerich==0.8.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncclick==8.1.8
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
dictdiffer==0.9.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
isort==5.13.2
jinja2==3.1.5
loguru==0.7.3
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.14
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.5
pydantic-core==2.27.2
pydantic-settings==2.7.1
pygments==2.19.1
pyjwt==2.10.1
pypika-tortoise==0.3.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyyaml==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
ruff==0.9.1
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.23.0
typer==0.15.1
typing-extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0; sys_platform != 'win32'
watchfiles==1.0.4
websockets==14.1
# 爬虫依赖
drissionpage==4.1.0.16
beautifulsoup4==4.12.3
